split to separate binaries
