+++ include/DroneState.hpp
@@ -33,9 +33,12 @@ public:
   DroneState &operator=(DroneState const &rhs);
   ~DroneState();
 
+  bool is_armed() const;
   std::string get_state() const;
+
   bool send_state(int serial_port);
   ssize_t send(int serial_port, std::string const &msg);
+  bool startup(int serial_port);
   std::string arm();
   std::string disarm();
   std::string setpoint(SetPoint const &point);
