+++ src/DroneState.cpp
@@ -1,11 +1,14 @@
 #include <iostream>
 #include <sstream>
-#include <unistd.h> // write()
+#include <string>
+
+#include <unistd.h>
 
 #include "DroneState.hpp"
 
 std::ostream &operator<<(std::ostream &os, SetPoint &p) {
-  os << "{'setpoint': [" << p.x << "," << p.y  << "," << p.z << "]}" << std::endl;
+  os << "{'setpoint': [" << p.x << "," << p.y << "," << p.z << "]}"
+     << std::endl;
   return os;
 }
 
@@ -26,6 +29,10 @@ DroneState &DroneState::operator=(DroneState const &rhs) {
 }
 
 DroneState::~DroneState() {}
+
+bool DroneState::is_armed() const {
+  return this->armed;
+}
 
 std::string DroneState::get_state() const {
   std::ostringstream ss;
@@ -55,7 +62,8 @@ bool DroneState::send_state(int serial_port) {
 }
 
 ssize_t DroneState::send(int serial_port, std::string const &msg) {
-  std::cout << "Port: " << serial_port << " msg: " << msg << " Len: " << msg.length() << std::endl;
+  // std::cout << "Port: " << serial_port << " msg: " << msg
+  //           << " Len: " << msg.length() << std::endl;
   return write(serial_port, msg.c_str(), msg.length());
 }
 
@@ -65,33 +73,48 @@ std::string DroneState::arm() {
   return std::string("{'armed':true}");
 }
 
+// Startup sequence requires throttling down to safe margin
+// to allow arming the drone
+bool DroneState::startup(int serial_port) {
+  std::string str(std::to_string(this->index));
+  str.append(this->trim(0, 0, -800, 0));
+  if (this->send(serial_port, str) < 0) {
+    std::cerr << "Failed to send throttle down signal" << std::endl;
+    return false;
+  }
+  str = std::to_string(this->index);
+  str.append(this->arm());
+  sleep(1);
+  if (this->send(serial_port, str) < 0) {
+    std::cerr << "Failed to send arming signal" << std::endl;
+    return false;
+  }
+  return true;
+}
+
 std::string DroneState::disarm() {
   this->armed = false;
   this->json["armed"] = "false";
   return std::string("{'armed':false}");
 }
 
-std::string DroneState::setpoint(SetPoint const &point)
-{
+std::string DroneState::setpoint(SetPoint const &point) {
   std::stringstream ss;
   ss << "{'setpoint':[" << point.x << "," << point.y << "," << point.z << "]}";
   this->json["setpoint"] = ss.str();
   return ss.str();
 }
 
-std::string DroneState::setpos(float x, float y, float z, float yaw)
-{
+std::string DroneState::setpos(float x, float y, float z, float yaw) {
   std::stringstream ss;
   ss << "{'pos':[" << x << "," << y << "," << z << "," << yaw << "]}";
   this->json["pos"] = ss.str();
   return ss.str();
 }
 
-std::string DroneState::trim(float x, float y, float z, float yaw)
-{
+std::string DroneState::trim(float x, float y, float z, float yaw) {
   std::stringstream ss;
   ss << "{'trim':[" << x << "," << y << "," << z << "," << yaw << "]}";
   this->json["trim"] = ss.str();
   return ss.str();
 }
-
